* dcrt0.cc (dll_crt0_1): Initialize thread and debug stuff before handling exec/fork. * dtable.cc (dtable::fixup_after_exec): Always clear out the read ahead buffer whether closing or adjusting. * path.cc (chdir): Avoid a compiler warning.

+++ winsup/cygwin/path.cc
@@ -2507,7 +2507,7 @@ chdir (const char *dir)
   else
     p++;
 
-  int len = strlen (p);
+  size_t len = strlen (p);
   if (len > 2 && strspn (p, ".") == len)
     {
       set_errno (ENOENT);
